rename module to keep naming convention
