+++ internal/log_analysis/log_processor/parsers/snapshotlogs/resource.go
@@ -1,4 +1,4 @@
-package snapshots
+package snapshotlogs
 
 /**
  * Panther is a Cloud-Native SIEM for the Modern Security Team.
